requirements.txt
# Core dependencies
torch>=1.12.0
gymnasium>=0.26.0
stable-baselines3>=2.0.0
isaacgym>=1.0.0

# Numerical and scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Visualization and plotting
matplotlib>=3.5.0
tensorboard>=2.8.0

# Robotics and planning
pybullet>=3.2.0  # For kinematics validation (optional)

# Utilities
tqdm>=4.62.0
PyYAML>=6.0
opencv-python>=4.5.0  # For visualization (optional)

# Development and testing
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0